DMD-452 - move tests
